Prepare 0.0.1 docs

+++ include/misc/tools_stats.hpp
@@ -68,15 +68,11 @@ namespace tools_stats
     };
     //! @}
 
-    Eigen::MatrixXd simulate_uniform(int n, int d);
-    Eigen::VectorXd to_pseudo_obs_1d(
-            Eigen::VectorXd x,
-            std::string ties_method = "average"
-    );
-    Eigen::MatrixXd to_pseudo_obs(
-            Eigen::MatrixXd x,
-            std::string ties_method = "average"
-    );
+    Eigen::MatrixXd simulate_uniform(size_t n, size_t d);
+    Eigen::VectorXd to_pseudo_obs_1d(Eigen::VectorXd x,
+                                     std::string ties_method = "average");
+    Eigen::MatrixXd to_pseudo_obs(Eigen::MatrixXd x,
+                                  std::string ties_method = "average");
 
     double pairwise_hoeffd(Eigen::Matrix<double, Eigen::Dynamic, 2> x);
     double pairwise_ktau(Eigen::Matrix<double, Eigen::Dynamic, 2>& u);
